MVP is ready

+++ src/modules/Meeting/index.js
@@ -1,3 +1,4 @@
+import { isAdmin } from '../../data/tableData';
 import './Meeeting.scss';
 
 class Meeting {
@@ -12,23 +13,26 @@ class Meeting {
     this.span = document.createElement('span');
     this.span.className = 'meeting';
     this.span.textContent = this.meeting.title;
-    this.span.setAttribute('draggable', true);
-
-    const deleteBtn = document.createElement('i');
-
-    deleteBtn.className = 'fa fa-times danger';
-    deleteBtn.onclick = () => {
-      this.onDelete(this.meeting);
-    };
-    this.span.appendChild(deleteBtn);
 
     this.span.setAttribute('data-id', this.meeting.id);
     this.span.id = this.meeting.id;
-
-    this.span.ondragstart = (e) => {
-      e.dataTransfer.setData('text/plain', e.target.id);
-      e.currentTarget.style.backgroundColor = 'yellow';
-    };
+    if (isAdmin) {
+      this.span.setAttribute('draggable', true);
+      this.span.style.cursor = 'pointer';
+      const deleteBtn = document.createElement('i');
+      deleteBtn.className = 'danger';
+      deleteBtn.className = 'fa fa-times danger';
+      deleteBtn.onclick = () => {
+        this.onDelete(this.meeting);
+      };
+      this.span.appendChild(deleteBtn);
+      this.span.ondragstart = (e) => {
+        e.dataTransfer.setData('text/plain', e.target.id);
+        e.currentTarget.style.backgroundColor = 'yellow';
+      };
+    } else {
+      this.span.style.justifyContent = 'center';
+    }
     this.timeSlot.appendChild(this.span);
   }
 }
